% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{is_collection}
\alias{is_collection}
\title{Check if directory is a webrarian collection}
\usage{
is_collection(path = ".")
}
\arguments{
\item{path}{Path to check}
}
\value{
\code{TRUE} if the directory contains a \verb{_webrarian.yml} file.
}
\description{
Check if directory is a webrarian collection
}
\examples{
is_collection(collection_example("basic"))

is_collection(tempdir())
}
