% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{print.webrarian_brand}
\alias{print.webrarian_brand}
\title{Print method for webrarian_brand}
\usage{
\method{print}{webrarian_brand}(x, ...)
}
\arguments{
\item{x}{A \code{webrarian_brand} object.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for webrarian_brand
}
