% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.webrarian_config}
\alias{print.webrarian_config}
\title{Print method for webrarian_config}
\usage{
\method{print}{webrarian_config}(x, ...)
}
\arguments{
\item{x}{A \code{webrarian_config} object.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for webrarian_config
}
