% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{print.webrarian_files}
\alias{print.webrarian_files}
\title{Print method for webrarian_files}
\usage{
\method{print}{webrarian_files}(x, ...)
}
\arguments{
\item{x}{A \code{webrarian_files} object from \code{\link[=collection_files]{collection_files()}}.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for webrarian_files
}
