% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{print.webrarian_packages}
\alias{print.webrarian_packages}
\title{Print method for webrarian_packages}
\usage{
\method{print}{webrarian_packages}(x, ...)
}
\arguments{
\item{x}{A \code{webrarian_packages} object.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for webrarian_packages
}
