% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{print.webrarian_server}
\alias{print.webrarian_server}
\title{Print a reading room}
\usage{
\method{print}{webrarian_server}(x, ...)
}
\arguments{
\item{x}{A reading room from \code{\link[=reading_room]{reading_room()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a reading room
}
